Terrible
The program runs fine, crafting hasn't been put in yet but it's wonderful... it's also abandoned because while it runs fine people look at the code and say "Oh my god", "Oh no", "How does this even work?", "That's not a thing you can do in Python", and so on. It's also full of spaghetti code. All problems I'm hoping to fix in the second go.