the buying&selling part
does the buying and selling, contains all the items, has code for checking if things are sell-able but doesn't run that part because player only allows sell-able items to be sold. Was planning on figuring out how to get that to work but will have to do it in the next go.